requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
absl-py=1.2.0=pypi_0
affine=2.3.1=pypi_0
aiohttp=3.8.1=py38h7f8727e_1
aiosignal=1.2.0=pyhd3eb1b0_0
asttokens=2.0.8=pypi_0
astunparse=1.6.3=pypi_0
async-timeout=4.0.2=py38h06a4308_0
attrs=22.1.0=py38h06a4308_0
backcall=0.2.0=pypi_0
blas=1.0=mkl
blinker=1.4=py38h06a4308_0
brotli=1.0.9=he6710b0_2
brotlipy=0.7.0=py38h27cfd23_1003
bzip2=1.0.8=h7b6447c_0
c-ares=1.18.1=h7f8727e_0
ca-certificates=2022.10.11=h06a4308_0
cachetools=4.2.2=pyhd3eb1b0_0
certifi=2022.12.7=py38h06a4308_0
cffi=1.15.0=py38hd667e15_1
chardet=3.0.4=pypi_0
charset-normalizer=2.0.12=pypi_0
click=8.1.3=pypi_0
click-plugins=1.1.1=pypi_0
cligj=0.7.2=pypi_0
cloudpickle=2.0.0=pyhd3eb1b0_0
cryptography=36.0.0=py38h9ce1e76_0
cuda=11.6.2=0
cuda-cccl=11.6.55=hf6102b2_0
cuda-command-line-tools=11.6.2=0
cuda-compiler=11.6.2=0
cuda-cudart=11.6.55=he381448_0
cuda-cudart-dev=11.6.55=h42ad0f4_0
cuda-cuobjdump=11.6.124=h2eeebcb_0
cuda-cupti=11.6.124=h86345e5_0
cuda-cuxxfilt=11.6.124=hecbf4f6_0
cuda-driver-dev=11.6.55=0
cuda-gdb=11.8.86=0
cuda-libraries=11.6.2=0
cuda-libraries-dev=11.6.2=0
cuda-memcheck=11.8.86=0
cuda-nsight=11.8.86=0
cuda-nsight-compute=11.8.0=0
cuda-nvcc=11.6.124=hbba6d2d_0
cuda-nvdisasm=11.8.86=0
cuda-nvml-dev=11.6.55=haa9ef22_0
cuda-nvprof=11.8.87=0
cuda-nvprune=11.6.124=he22ec0a_0
cuda-nvrtc=11.6.124=h020bade_0
cuda-nvrtc-dev=11.6.124=h249d397_0
cuda-nvtx=11.6.124=h0630a44_0
cuda-nvvp=11.8.87=0
cuda-runtime=11.6.2=0
cuda-samples=11.6.101=h8efea70_0
cuda-sanitizer-api=11.8.86=0
cuda-toolkit=11.6.2=0
cuda-tools=11.6.2=0
cuda-visual-tools=11.6.2=0
cudatoolkit=10.2.89=hfd86e86_1
cycler=0.11.0=pyhd3eb1b0_0
cytoolz=0.11.0=py38h7b6447c_0
dask-core=2022.5.0=py38h06a4308_0
dataclasses=0.8=pyh6d0b6a4_7
dbus=1.13.18=hb2f20db_0
decorator=5.1.1=pyhd3eb1b0_0
deprecated=1.2.13=pypi_0
dgl-cuda10.2=0.8.1=py38_0
direct=1.0.3=pypi_0
docker-pycreds=0.4.0=pypi_0
executing=1.0.0=pypi_0
expat=2.4.4=h295c915_0
ffmpeg=4.3=hf484d3e_0
flatbuffers=1.12=pypi_0
fontconfig=2.13.1=h6c09931_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.11.0=h70c0345_0
frozenlist=1.2.0=py38h7f8727e_0
fsspec=2022.3.0=py38h06a4308_0
gast=0.4.0=pypi_0
gds-tools=1.4.0.31=0
giflib=5.2.1=h7b6447c_0
gitdb=4.0.9=pypi_0
gitpython=3.1.29=pypi_0
glib=2.69.1=h4ff587b_1
gmp=6.2.1=h2531618_2
gnutls=3.6.15=he1e5248_0
google-auth=2.6.0=pyhd3eb1b0_0
google-auth-oauthlib=0.4.4=pyhd3eb1b0_0
google-pasta=0.2.0=pypi_0
grpcio=1.42.0=py38hce63b2e_0
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
h5py=3.3.0=pypi_0
humanize=4.4.0=pypi_0
icu=58.2=he6710b0_3
idna=2.8=pypi_0
ignite=0.4.10=py_0
image-similarity-measures=0.3.5=pypi_0
imageio=2.9.0=pyhd3eb1b0_0
imageio-ffmpeg=0.4.7=pypi_0
importlib-metadata=4.11.3=py38h06a4308_0
intel-openmp=2021.4.0=h06a4308_3561
ipdb=0.13.9=pypi_0
ipython=8.5.0=pypi_0
ismrmrd=1.9.8=pypi_0
ismrmrd-python-tools=0.3=pypi_0
jedi=0.18.1=pypi_0
jpeg=9d=h7f8727e_0
jsonpatch=1.32=pypi_0
jsonpointer=2.3=pypi_0
keras=2.9.0=pypi_0
keras-preprocessing=1.1.2=pypi_0
kiwisolver=1.4.2=py38h295c915_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.35.1=h7274673_9
libclang=14.0.1=pypi_0
libcublas=11.11.3.6=0
libcublas-dev=11.11.3.6=0
libcufft=10.9.0.58=0
libcufft-dev=10.9.0.58=0
libcufile=1.4.0.31=0
libcufile-dev=1.4.0.31=0
libcurand=10.3.0.86=0
libcurand-dev=10.3.0.86=0
libcusolver=11.4.1.48=0
libcusolver-dev=11.4.1.48=0
libcusparse=11.7.5.86=0
libcusparse-dev=11.7.5.86=0
libffi=3.3=he6710b0_2
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=7.5.0=ha8ba4b0_17
libgfortran4=7.5.0=ha8ba4b0_17
libgomp=11.2.0=h1234567_1
libiconv=1.15=h63c8f33_5
libidn2=2.3.2=h7f8727e_0
libnpp=11.8.0.86=0
libnpp-dev=11.8.0.86=0
libnvjpeg=11.9.0.86=0
libnvjpeg-dev=11.9.0.86=0
libpng=1.6.37=hbc83047_0
libprotobuf=3.20.1=h4ff587b_0
libstdcxx-ng=9.3.0=hd4cf53a_17
libtasn1=4.16.0=h27cfd23_0
libtiff=4.2.0=h85742a9_0
libunistring=0.9.10=h27cfd23_0
libuuid=1.0.3=h7f8727e_2
libuv=1.40.0=h7b6447c_0
libwebp=1.2.2=h55f646e_0
libwebp-base=1.2.2=h7f8727e_0
libxcb=1.15=h7f8727e_0
libxml2=2.9.12=h03d6c58_0
locket=1.0.0=py38h06a4308_0
lz4-c=1.9.3=h295c915_1
markdown=3.3.4=py38h06a4308_0
matplotlib=3.5.1=py38h06a4308_1
matplotlib-base=3.5.1=py38ha18d171_1
matplotlib-inline=0.1.6=pypi_0
metrics=0.3.3=pypi_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py38h7f8727e_0
mkl_fft=1.3.1=py38hd3c417c_0
mkl_random=1.2.2=py38h51133e4_0
moviepy=1.0.3=pypi_0
multidict=6.0.2=py38h5eee18b_0
munkres=1.1.4=py_0
ncurses=6.3=h7f8727e_2
nettle=3.7.3=hbbd107a_1
networkx=2.7.1=pyhd3eb1b0_0
nibabel=4.0.2=pypi_0
nose=1.3.7=pypi_0
nsight-compute=2022.3.0.22=0
numpy=1.21.0=pypi_0
numpy-base=1.21.2=py38h79a1101_0
oauthlib=3.2.0=pypi_0
omegaconf=2.1.1=pypi_0
opencv-python=4.6.0.66=pypi_0
openh264=2.1.1=h4ff587b_0
openssl=1.1.1s=h7f8727e_0
opt-einsum=3.3.0=pypi_0
packaging=21.3=pyhd3eb1b0_0
pandas=1.4.2=pypi_0
parso=0.8.3=pypi_0
partd=1.2.0=pyhd3eb1b0_1
pathlib2=2.3.7.post1=pypi_0
pathspec=0.5.5=pypi_0
pathtools=0.1.2=pypi_0
pcre=8.45=h295c915_0
pexpect=4.8.0=pypi_0
phasepack=1.5=pypi_0
pickleshare=0.7.5=pypi_0
pillow=9.0.1=py38h22f2fdc_0
pip=22.2.2=pypi_0
proglog=0.1.9=pypi_0
promise=2.3=pypi_0
prompt-toolkit=3.0.31=pypi_0
protobuf=3.20.1=pypi_0
psutil=5.9.3=pypi_0
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.2=pypi_0
pyasn1=0.4.8=pyhd3eb1b0_0
pyasn1-modules=0.2.8=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pyfftw=0.13.0=pypi_0
pygments=2.13.0=pypi_0
pyjwt=2.4.0=py38h06a4308_0
pynufft=2022.2.3=pypi_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.4=pyhd3eb1b0_0
pyprof2calltree=1.4.5=pypi_0
pyqt=5.9.2=py38h05f1152_4
pysocks=1.7.1=py38h06a4308_0
python=3.8.13=h12debd9_0
python-dateutil=2.8.2=pyhd3eb1b0_0
pytorch=1.13.0=py3.8_cuda11.6_cudnn8.3.2_0
pytorch-complex=0.0.8=pypi_0
pytorch-cuda=11.6=h867d48c_0
pytorch-mutex=1.0=cuda
pytorch-ssim=0.1=pypi_0
pytz=2022.1=pypi_0
pywavelets=1.3.0=py38h7f8727e_0
pyxb=1.2.6=pypi_0
pyyaml=6.0=py38h7f8727e_1
pyzmq=23.2.1=pypi_0
qt=5.9.7=h5867ecd_1
rasterio=1.3.4=pypi_0
readline=8.1.2=h7f8727e_1
requests=2.27.1=pyhd3eb1b0_0
requests-oauthlib=1.3.0=py_0
rsa=4.7.2=pyhd3eb1b0_1
runstats=1.7.1=pypi_0
scikit-image=0.15.0=pypi_0
scikit-learn=1.1.2=pypi_0
scipy=1.4.1=pypi_0
sentry-sdk=1.10.1=pypi_0
setproctitle=1.3.2=pypi_0
setuptools=62.6.0=pypi_0
sewar=0.4.5=pypi_0
shortuuid=1.0.9=pypi_0
simpleitk=2.2.0=pypi_0
sip=4.19.13=py38h295c915_0
six=1.12.0=pypi_0
smmap=5.0.0=pypi_0
snuggs=1.4.7=pypi_0
sqlite=3.38.2=hc218d9a_0
stack-data=0.5.0=pypi_0
tensorboard=2.9.1=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=py38h06a4308_0
tensorboardx=2.1=pypi_0
tensorflow=2.9.1=pypi_0
tensorflow-estimator=2.9.0=pypi_0
tensorflow-io-gcs-filesystem=0.26.0=pypi_0
termcolor=1.1.0=pypi_0
tk=8.6.11=h1ccaba5_0
toolz=0.11.2=pyhd3eb1b0_0
torch=1.11.0=pypi_0
torch-complex=0.4.3=pypi_0
torchaudio=0.13.0=py38_cu116
torchfile=0.1.0=pypi_0
torchio=0.18.84=pypi_0
torchmetrics=0.11.0=pyhd8ed1ab_0
torchnet=0.0.4=pypi_0
torchvision=0.7.0=pypi_0
tornado=6.1=py38h27cfd23_0
tqdm=4.64.0=pypi_0
traitlets=5.4.0=pypi_0
tsp-solver2=0.3=pypi_0
typing_extensions=4.1.1=pyh06a4308_0
urllib3=1.26.12=pypi_0
visdom=0.1.8.9=pypi_0
wandb=0.13.4=pypi_0
wcwidth=0.2.5=pypi_0
websocket-client=1.4.1=pypi_0
werkzeug=2.0.3=pyhd3eb1b0_0
wheel=0.37.1=pypi_0
wrapt=1.14.1=pypi_0
xsdata=22.2=pypi_0
xz=5.2.5=h7b6447c_0
yaml=0.2.5=h7b6447c_0
yarl=1.8.1=py38h5eee18b_0
zipp=3.8.0=py38h06a4308_0
zlib=1.2.13=h5eee18b_0
zstd=1.4.9=haebb681_0
